dlx: init: Configure cpu-boost's plug_boost
Using only a 1 ms boost for dlx because the dlx mpdecision is
very aggressive about bringing cores online.

The only purpose of this boost on dlx is to ensure that the core
doesn't come online at very low frequency and then take a little
time for the governor to correctly identify the load on it.

Change-Id: Ic56d375b91d35271cb593499ea172f877326a7fe